+++ ggc/core/Batch.java
@@ -15,7 +15,7 @@ public class Batch {
     }
 
     public String toString(){
-        return _product.getId() + "|" + _partner.getId() + "|" + _price + "|" + _quantity;
+        return _product.getId() + "|" + _partner.getId() + "|" + (int)_price + "|" + _quantity;
     }
 
     public int getQuantity(){
